% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hourly_time_series}
\alias{hourly_time_series}
\title{Hourly rainfall time series}
\format{A data frame with columns:
\describe{
 \item{Date}{The time and day of the rainfall pulse}
 \item{Rainfall.depth}{The depth of rainfall in each time step}
}}
\source{
Artificial data, it does not come from any source.
}
\usage{
hourly_time_series
}
\description{
An artificial hourly rainfall dataset to test the functions of the package.
}
\examples{
\dontrun{
 hourly_time_series
}
}
\keyword{datasets}
